normi finit, go tester ouuuhouu

+++ ft_calloc.c
@@ -6,7 +6,7 @@
 /*   By: nbenhado <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/26 16:15:38 by nbenhado          #+#    #+#             */
-/*   Updated: 2021/11/26 16:40:32 by nbenhado         ###   ########.fr       */
+/*   Updated: 2021/11/26 18:26:09 by nbenhado         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 void	*ft_calloc( size_t elementCount, size_t elementSize )
 {
-	int		i;
+	size_t	i;
 	char	*block;
 
 	block = malloc(elementSize * elementCount);
